+++ src/app/api/menu-items/route.js
@@ -1,46 +1,62 @@
 import { MenuItem } from "@/app/models/MenuItem";
-import mongoose from "mongoose";
 import { NextResponse } from "next/server";
+import { isAdmin } from "../auth/[...nextauth]/route";
+import { connectDB } from "@/libs/connectDB";
 
 export async function POST(req) {
   try {
-    await mongoose.connect(process.env.MONGO_URL);
+    await connectDB();
+    if (!(await isAdmin())) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 403 });
+    }
     const data = await req.json();
     const createdMenuItem = await MenuItem.create(data);
     return NextResponse.json(createdMenuItem);
   } catch (error) {
     console.error("Error creating item:", error.message);
-    return NextResponse.json(error.message, { status: 500 });
+    return NextResponse.json({ error: error.message }, { status: 500 });
   }
 }
 
 export async function PUT(req) {
   try {
-    await mongoose.connect(process.env.MONGO_URL);
+    await connectDB();
+    if (!(await isAdmin())) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 403 });
+    }
     const { _id, ...data } = await req.json();
-    const updatedItem = await MenuItem.findByIdAndUpdate(_id, data);
+    const updatedItem = await MenuItem.findByIdAndUpdate(_id, data, {
+      new: true,
+    });
     return NextResponse.json(updatedItem);
   } catch (error) {
     console.error("Error updating item:", error.message);
-    return NextResponse.json(error.message, { status: 500 });
+    return NextResponse.json({ error: error.message }, { status: 500 });
   }
 }
 
 export async function GET() {
   try {
-    await mongoose.connect(process.env.MONGO_URL);
-    return NextResponse.json(await MenuItem.find());
-  } catch (error) {}
+    await connectDB();
+    const items = await MenuItem.find();
+    return NextResponse.json(items);
+  } catch (error) {
+    console.error("Error fetching items:", error.message);
+    return NextResponse.json({ error: error.message }, { status: 500 });
+  }
 }
 
 export async function DELETE(req) {
   try {
-    await mongoose.connect(process.env.MONGO_URL);
+    await connectDB();
+    if (!(await isAdmin())) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 403 });
+    }
     const _id = req.nextUrl.searchParams.get("_id");
     await MenuItem.deleteOne({ _id });
-    return NextResponse.json(true);
+    return NextResponse.json({ success: true });
   } catch (error) {
-    console.error("Error deleting menu item:", error);
+    console.error("Error deleting menu item:", error.message);
     return NextResponse.json(
       { error: "Error deleting menu item" },
       { status: 500 }
